Don't use Timeout.timeout in Server#boot
I expect this to be equivalent since it rases the same error and has the
same pattern of sleeps.

I don't know of any problems caused by using timeout here, instead I've
converted this for consistency and so that that there are no longer any
uses of `Timeout.timeout` within the codebase.